move plugin headers out of interface to avoid circular dependency
