+++ CalibTracker/SiStripCommon/plugins/BuildFile.xml
@@ -1,6 +1,7 @@
 <use name="DataFormats/L1GlobalTrigger"/>
 <use name="TrackingTools/TrackFitters"/>
 <use name="CalibTracker/SiStripCommon"/>
+<use name="CalibFormats/SiStripObjects"/>
 <use name="Geometry/TrackerGeometryBuilder"/>
 <use name="CondFormats/DataRecord"/>
 <use name="CondFormats/RunInfo"/>
