+++ CalibTracker/SiStripCommon/plugins/ShallowExampleProducer.h
