+++ CalibTracker/SiStripCommon/plugins/ShallowRechitClustersProducer.h
