+++ CalibTracker/SiStripCommon/plugins/ShallowTracksProducer.cc
@@ -1,4 +1,4 @@
-#include "CalibTracker/SiStripCommon/interface/ShallowTracksProducer.h"
+#include "ShallowTracksProducer.h"
 
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
